[2019-01-04] Money now rounded to 2 dp, mb/balance can display balances of other users
- Money given is now rounded to 2 d.p. and formatted like an actual currency
- mb/balance can now return balances of other users if you give it a user to search for

+++ MarbleBot/Modules/MoneyUser.cs
@@ -6,7 +6,7 @@ namespace MarbleBot.Modules
     {
         public string Name { get; set; }
         public string Discriminator { get; set; }
-        public ulong Money { get; set; }
+        public decimal Money { get; set; }
         public uint DailyStreak { get; set; }
         public DateTime LastDaily { get; set; }
         public DateTime LastRaceWin { get; set;}
